+++ common/src/main/java/earth/terrarium/botarium/util/BlockProvider.java
@@ -1,4 +1,0 @@
-package earth.terrarium.botarium.util;
-
-public interface BlockProvider<T> extends Updatable, SnapshotProvider<T> {
-}