+++ common/src/main/java/earth/terrarium/botarium/util/ItemProvider.java
@@ -1,7 +1,0 @@
-package earth.terrarium.botarium.util;
-
-import net.minecraft.world.item.ItemStack;
-
-public interface ItemProvider<T> extends Updatable, SnapshotProvider<T> {
-    ItemStack getContainerItem();
-}