+++ common/src/main/java/earth/terrarium/botarium/util/SnapshotProvider.java
@@ -1,7 +1,0 @@
-package earth.terrarium.botarium.util;
-
-public interface SnapshotProvider<T> {
-    T createSnapshot();
-
-    void loadSnapshot(T snapshot);
-}