+++ fabric/src/main/java/earth/terrarium/botarium/fabric/extensions/ContainerApiImpl.java
@@ -1,7 +1,7 @@
 package earth.terrarium.botarium.fabric.extensions;
 
 import earth.terrarium.botarium.common.generic.base.BlockContainerLookup;
-import earth.terrarium.botarium.common.generic.ContainerApi;
+import earth.terrarium.botarium.common.generic.LookupApi;
 import earth.terrarium.botarium.common.generic.base.EntityContainerLookup;
 import earth.terrarium.botarium.common.generic.base.ItemContainerLookup;
 import earth.terrarium.botarium.fabric.generic.FabricBlockContainerLookup;
@@ -9,11 +9,9 @@ import earth.terrarium.botarium.fabric.generic.FabricEntityContainerLookup;
 import earth.terrarium.botarium.fabric.generic.FabricItemContainerLookup;
 import net.minecraft.resources.ResourceLocation;
 import net.msrandom.extensions.annotations.ClassExtension;
-import net.msrandom.extensions.annotations.ImplementedByExtension;
 import net.msrandom.extensions.annotations.ImplementsBaseElement;
-import org.apache.commons.lang3.NotImplementedException;
 
-@ClassExtension(ContainerApi.class)
+@ClassExtension(LookupApi.class)
 public class ContainerApiImpl {
     @ImplementsBaseElement
     public static <T, C> BlockContainerLookup<T, C> createBlockLookup(ResourceLocation name, Class<T> typeClass, Class<C> contextClass) {