+++ fabric/src/main/java/earth/terrarium/botarium/fabric/extensions/PlatformItemContainerImpl.java
@@ -1,7 +1,7 @@
 package earth.terrarium.botarium.fabric.extensions;
 
-import earth.terrarium.botarium.common.item.ItemContainer;
-import earth.terrarium.botarium.fabric.item.FabricItemContainer;
+import earth.terrarium.botarium.common.item.base.ItemContainer;
+import earth.terrarium.botarium.fabric.item.PlatformItemContainer;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.world.level.Level;
@@ -16,6 +16,6 @@ public class PlatformItemContainerImpl {
 
     @ImplementsBaseElement
     static ItemContainer of(Level level, BlockPos pos, BlockState state, @Nullable BlockEntity entity, @Nullable Direction direction) {
-        return FabricItemContainer.of(level, pos, state, entity, direction);
+        return PlatformItemContainer.of(level, pos, state, entity, direction);
     }
 }